# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "social api",
        "title": "Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/building": {
            "put": {
                "description": "update building",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update building",
                "operationId": "update-building",
                "parameters": [
                    {
                        "description": "updatable building",
                        "name": "building",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/building.Building"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "create building",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create building",
                "operationId": "create-building",
                "parameters": [
                    {
                        "description": "creatable building",
                        "name": "building",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/building.Building"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/building.Building"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    }
                }
            }
        },
        "/building/{id}": {
            "get": {
                "description": "get building by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get building by id",
                "operationId": "get-building-by-id",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/building.Building"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete building by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete building by id",
                "operationId": "delete-building-by-id",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    }
                }
            }
        },
...
        "/buildings": {
            "get": {
                "description": "get buildings by params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get buildings",
                "operationId": "get-buildings",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "with organization",
                        "name": "with_organization",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/building.BuildingsList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_error.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_error.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "building.Building": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "coords": {
                    "$ref": "#/definitions/coords.Coords"
                },
                "id": {
                    "type": "integer"
                },
                "organizations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/organization.Organization"
                    }
                }
            }
        },
        "building.BuildingsList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/building.Building"
                    }
                }
            }
        },
...
        "coords.Coords": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "organization.Organization": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/organization_phone.OrganizationPhone"
                    }
                }
            }
        },
...
        "organization_phone.OrganizationPhone": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "integer"
                }
            }
        }
    }
}